Avoid creation of thousands of Ramda thunks
Here's the thing: as much as I love the curried goodness of Ramda and
the Haskell-y feel it provides, no JS runtime is optimized for the
creation of thousands of functions in a hot path.

Using it in the first passes Lev.levenshtein and BKTree.searchWords has
given me confidence when implementing their property tests, but
profiling the code has shown how costly the abstraction is. Replacing Ramda
functions with native equivalents has yielded a 5-8x difference when
searching a ~260k words space with a threshold of 3.