Fix pipeline order in bk_tree_test
The property for "keeps every word that's inserted" sometimes failed
when the set of new words to add included the same word in difference
case styles (for instance, `["u", "U"]`). Even though the test was
processing that input to get only unique lowercase items, it was doing
it in the *wrong* order: uniq'ing and then lowercasing (which, in the
exampel input, would result in `["u", "u"]`).